requirements.txt
imagehash
progressbar2